ResourceManager Image Filter Selection
Updated the ResourceManager class to require a filter selection before
any resources are loaded, also created an overload so filter can
optionally be changed on a per image basis. Because filter is chosen
when the image is read off disk this setting is required to be set
before any files are accessed.